+++ src/aks-preview/HISTORY.rst
@@ -11,10 +11,13 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
+* Add `--nodepool-initialization-taints` to `az aks create` and `az aks update`.
 
 3.0.0b2
 +++++++
-* Add `--nodepool-initialization-taints` to `az aks create` and `az aks update`.
+* Add `--sku` to the `az aks create` command.
+* Add `--sku` to the `az aks update` command.
+* Support cluster service health probe mode by `--cluster-service-load-balancer-health-probe-mode {Shared, Servicenodeport}`
 
 3.0.0b1
 +++++++